# service/src/main/resources/api/service_openapi.yaml
openapi: 3.0.3
info:
  title: Tanagra Service API
  description: github.com/DataBiosphere/tanagra
  version: 0.0.0
servers:
  - url: /
security:
  - bearerAuth: []
  - authorization: [openid, email, profile]
paths:
...
  "/v2/studies/{studyId}/cohorts/{cohortId}":
    parameters:
      - $ref: '#/components/parameters/StudyIdV2'
      - $ref: '#/components/parameters/CohortIdV2'
    get:
      summary: Get an existing cohort
      operationId: getCohort
      tags: [CohortsV2]
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CohortV2'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/ServerError'
    patch:
      summary: Update an existing cohort
      operationId: updateCohort
      tags: [CohortsV2]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CohortUpdateInfoV2'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CohortV2'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/ServerError'
    delete:
      summary: Delete a cohort
      operationId: deleteCohort
      tags: [CohortsV2]
      responses:
        204:
          description: OK
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/ServerError'
  "/v2/studies/{studyId}/cohorts/{cohortId}/reviews":
    parameters:
      - $ref: '#/components/parameters/StudyIdV2'
      - $ref: '#/components/parameters/CohortIdV2'
    get:
      parameters:
        - $ref: "#/components/parameters/OffsetV2"
        - $ref: "#/components/parameters/LimitV2"
      summary: List all reviews for a cohort
      operationId: listReviews
      tags: [ReviewsV2]
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewListV2"
        500:
          $ref: "#/components/responses/ServerError"
    post:
      summary: Create a new review
      operationId: createReview
      tags: [ReviewsV2]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewCreateInfoV2"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewV2"
        500:
          $ref: "#/components/responses/ServerError"
  "/v2/studies/{studyId}/cohorts/{cohortId}/reviews/{reviewId}":
    parameters:
      - $ref: '#/components/parameters/StudyIdV2'
      - $ref: '#/components/parameters/CohortIdV2'
      - $ref: '#/components/parameters/ReviewIdV2'
    get:
      summary: Get an existing review
      operationId: getReview
      tags: [ReviewsV2]
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewV2'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/ServerError'
    patch:
      summary: Update an existing review
      operationId: updateReview
      tags: [ReviewsV2]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewUpdateInfoV2'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewV2'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/ServerError'
    delete:
      summary: Delete a review
      operationId: deleteReview
      tags: [ReviewsV2]
      responses:
        204:
          description: OK
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/ServerError'
  "/v2/studies/{studyId}/cohorts/{cohortId}/reviews/{reviewId}/instances":
    parameters:
      - $ref: '#/components/parameters/StudyIdV2'
      - $ref: '#/components/parameters/CohortIdV2'
      - $ref: '#/components/parameters/ReviewIdV2'
    post:
      summary: List primary entity instances in a review and any associated annotation values
      operationId: listReviewInstancesAndAnnotations
      tags: [ReviewsV2]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewQueryV2'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewInstanceListV2"
        500:
          $ref: "#/components/responses/ServerError"
  "/v2/studies/{studyId}/cohorts/{cohortId}/reviews/{reviewId}/counts":
    parameters:
      - $ref: '#/components/parameters/StudyIdV2'
      - $ref: '#/components/parameters/CohortIdV2'
      - $ref: '#/components/parameters/ReviewIdV2'
    post:
      summary: Count primary entity instances in a review
      operationId: countReviewInstances
      tags: [ReviewsV2]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewCountQueryV2'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstanceCountListV2"
        500:
          $ref: "#/components/responses/ServerError"
...
  "/v2/studies/{studyId}/cohorts/{cohortId}/annotations/{annotationId}":
    parameters:
      - $ref: '#/components/parameters/StudyIdV2'
      - $ref: '#/components/parameters/CohortIdV2'
      - $ref: '#/components/parameters/AnnotationIdV2'
    get:
      summary: Get an existing annotation
      operationId: getAnnotation
      tags: [AnnotationsV2]
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnnotationV2'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/ServerError'
    patch:
      summary: Update an existing annotation
      operationId: updateAnnotation
      tags: [AnnotationsV2]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnnotationUpdateInfoV2'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnnotationV2'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/ServerError'
    delete:
      summary: Delete an annotation
      operationId: deleteAnnotation
      tags: [AnnotationsV2]
      responses:
        204:
          description: OK
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/ServerError'
...
components:
  parameters:
...
    StudyIdV2:
      name: studyId
      in: path
      description: ID of the study
      required: true
      schema:
        type: string
    CohortIdV2:
      name: cohortId
      in: path
      description: ID of the cohort
      required: true
      schema:
        type: string
...
    ReviewIdV2:
      name: reviewId
      in: path
      description: ID of the review
      required: true
      schema:
        type: string
    AnnotationIdV2:
      name: annotationId
      in: path
      description: ID of the annotation
      required: true
      schema:
        type: string
...
    OffsetV2:
      name: offset
      in: query
      description: The number of items to skip before starting to collect the result set.
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    LimitV2:
      name: limit
      in: query
      description: The maximum number of items to return. Default 15
      required: false
      schema:
        type: integer
        minimum: 1
        default: 15
  responses:
    NotFound:
      description: Not found (or unauthorized)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorReportV2'
    ServerError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorReportV2'
  schemas:
...
    DataTypeV2:
      type: string
      enum: ["INT64", "STRING", "BOOLEAN", "DATE"]
...
    ReviewQueryV2:
      type: object
      description: Query for review instances and annotations
      properties:
        includeAttributes:
          description: Attributes to include in the returned instances
          type: array
          items:
            type: string
        entityFilter:
          $ref: "#/components/schemas/FilterV2"
        annotationFilter:
          $ref: "#/components/schemas/AnnotationFilterV2"
        orderBys:
          type: array
          description: Attributes or annotations, and direction to order the results by
          items:
            type: object
            description: Attribute or annotation and the direction
            properties:
              attribute:
                type: string
                description: Name of the attribute
              annotation:
                type: string
                description: ID of the annotation
              direction:
                $ref: "#/components/schemas/OrderByDirectionV2"
        limit:
          $ref: "#/components/schemas/LimitV2"
    ReviewCountQueryV2:
      type: object
      description: Count review instances
      properties:
        attributes:
          description: |
            Attributes to group by. One count will be returned for each possible combination of attribute values 
            (e.g. [gender, hair_color] will return counts for man+red, man+black, woman+red, woman+black).
          type: array
          items:
            type: string
    FilterV2:
      type: object
      description: Union of references to each filter type. Exactly one should be specified based on the filterType.
      properties:
        filterType:
          type: string
          enum: ["ATTRIBUTE", "TEXT", "HIERARCHY", "RELATIONSHIP", "BOOLEAN_LOGIC"]
        filterUnion:
          type: object
          properties:
            attributeFilter:
              $ref: "#/components/schemas/AttributeFilterV2"
            textFilter:
              $ref: "#/components/schemas/TextFilterV2"
            hierarchyFilter:
              $ref: "#/components/schemas/HierarchyFilterV2"
            relationshipFilter:
              $ref: "#/components/schemas/RelationshipFilterV2"
            booleanLogicFilter:
              $ref: "#/components/schemas/BooleanLogicFilterV2"
    AttributeFilterV2:
      type: object
      description: Filter on an attribute value (e.g. color=red)
      properties:
        attribute:
          type: string
        operator:
          $ref: "#/components/schemas/BinaryOperatorV2"
        value:
          $ref: "#/components/schemas/LiteralV2"
...
    AnnotationFilterV2:
      type: object
      description: Filter on an annotation value (e.g. reviewStatus=complete)
      properties:
        annotation:
          type: string
        operator:
          type: string
          enum: ["EQUALS", "NOT_EQUALS", "LESS_THAN", "GREATER_THAN"]
        value:
          $ref: "#/components/schemas/LiteralV2"
    OrderByDirectionV2:
      type: string
      description: Defaults to ascending
      enum: ["ASCENDING", "DESCENDING"]
    LimitV2:
      type: integer
      description: Maximum number of results to return. Defaults to 250.
      default: 250
...
    InstanceCountV2:
      type: object
      properties:
        count:
          type: integer
        attributes:
          description: A map of entity attribute names to their value for this group
          type: object
          additionalProperties:
            type: object
            nullable: true
            $ref: "#/components/schemas/ValueDisplayV2"
    InstanceCountListV2:
      type: object
      description: List of instance counts
      properties:
        sql:
          type: string
        instanceCounts:
          type: array
          items:
            $ref: "#/components/schemas/InstanceCountV2"
...
    CohortV2:
      type: object
      properties:
        id:
          description: ID of the cohort, immutable
          type: string
        underlayName:
          $ref: "#/components/schemas/UnderlayNameV2"
        displayName:
          $ref: "#/components/schemas/CohortDisplayNameV2"
        description:
          $ref: "#/components/schemas/CohortDescriptionV2"
        criteriaGroupSections:
          type: array
          description: Criteria group sections that define the entity filter
          items:
            $ref: "#/components/schemas/CriteriaGroupSectionV3"
        created:
          description: Timestamp of when the cohort was created
          type: string
          format: date-time
        createdBy:
          description: Email of user who created cohort
          type: string
        lastModified:
          description: Timestamp of when the cohort was last modified
          type: string
          format: date-time
      required:
        - id
        - underlayName
        - displayName
        - criteriaGroups
        - created
        - createdBy
        - lastModified
...
    CohortUpdateInfoV2:
      type: object
      properties:
        displayName:
          $ref: "#/components/schemas/CohortDisplayNameV2"
        description:
          $ref: "#/components/schemas/CohortDescriptionV2"
        criteriaGroupSections:
          type: array
          description: Criteria group sections that define the entity filter
          items:
            $ref: "#/components/schemas/CriteriaGroupSectionV3"
    CohortDisplayNameV2:
      type: string
      description: Human readable name of the cohort
    CohortDescriptionV2:
      type: string
      description: Description of the cohort
...
    UnderlayNameV2:
      type: string
      description: Name of the underlay, immutable
...
    CriteriaGroupSectionV3:
      type: object
      description: Section of criteria groups and boolean logic operators to combine them
      properties:
        id:
          type: string
          description: ID of the section, immutable
        displayName:
          type: string
          description: Name of the section
        criteriaGroups:
          type: array
          description: Set of criteria groups in the section
          items:
            $ref: "#/components/schemas/CriteriaGroupV3"
        operator:
          type: string
          description: Operator to use when combining the criteria groups in the section
          enum: ["AND", "OR"]
        excluded:
          type: boolean
          description: True to exclude the section, false to include it
      required:
        - id
        - displayName
        - criteriaGroups
        - operator
        - excluded
...
    ReviewV2:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/ReviewIdV2"
        displayName:
          $ref: "#/components/schemas/ReviewDisplayNameV2"
        description:
          $ref: "#/components/schemas/ReviewDescriptionV2"
        size:
          $ref: "#/components/schemas/ReviewSizeV2"
        cohort:
          $ref: "#/components/schemas/CohortV2"
        created:
          description: Timestamp of when the review was created
          type: string
          format: date-time
        createdBy:
          description: Email of user who created the review
          type: string
        lastModified:
          description: Timestamp of when the review was last modified
          type: string
          format: date-time
      required:
        - id
        - displayName
        - size
        - cohortRevision
        - created
        - createdBy
        - lastModified
    ReviewListV2:
      type: array
      items:
        $ref: "#/components/schemas/ReviewV2"
    ReviewCreateInfoV2:
      type: object
      properties:
        displayName:
          $ref: "#/components/schemas/ReviewDisplayNameV2"
        description:
          $ref: "#/components/schemas/ReviewDescriptionV2"
        size:
          $ref: "#/components/schemas/ReviewSizeV2"
        filter:
          # TODO: Remove this filter property from the service API once we fully support UI plugins, because the
          # plugin-specific backend code will generate it from the cohort definition. For now, the UI builds the
          # cohort filter and passes it to the backend when we create a review, so the backend can take a random
          # sample and preserve it.
          $ref: "#/components/schemas/FilterV2"
      required:
        - displayName
        - size
        - filter
    ReviewUpdateInfoV2:
      type: object
      properties:
        displayName:
          $ref: "#/components/schemas/ReviewDisplayNameV2"
        description:
          $ref: "#/components/schemas/ReviewDescriptionV2"
    ReviewIdV2:
      type: string
      description: ID of the review, immutable
    ReviewDisplayNameV2:
      type: string
      description: Human readable name of the review
    ReviewDescriptionV2:
      type: string
      description: Description of the review
    ReviewSizeV2:
      type: integer
      description: Number of primary entity instances included in the review
    AnnotationV2:
      type: object
      properties:
        id:
          description: ID of the annotation, immutable
          type: string
        displayName:
          $ref: "#/components/schemas/AnnotationDisplayNameV2"
        description:
          $ref: "#/components/schemas/AnnotationDescriptionV2"
        dataType:
          $ref: "#/components/schemas/DataTypeV2"
        enumVals:
          type: array
          description: |
            May only be set if dataType=STRING. Annotation value must be one of
            enumVals.
          items:
            type: string
      required:
        - id
        - displayName
        - dataType
...
    AnnotationUpdateInfoV2:
      type: object
      properties:
        displayName:
          $ref: "#/components/schemas/AnnotationDisplayNameV2"
        description:
          $ref: "#/components/schemas/AnnotationDescriptionV2"
    AnnotationDisplayNameV2:
      type: string
      description: Human readable name of the annotation
    AnnotationDescriptionV2:
      type: string
      description: Description of the annotation
...
    ReviewInstanceV2:
      type: object
      properties:
        attributes:
          description: |
            A map of entity attribute names to their value for this instance. The id attribute will always be included.
          type: object
          additionalProperties:
            type: object
            nullable: true
            $ref: "#/components/schemas/ValueDisplayV2"
        annotations:
          description: A map of annotation ids to their values for this instance in this review
          type: object
          additionalProperties:
            type: array
            nullable: true
            items:
              $ref: "#/components/schemas/AnnotationValueV2"
    ReviewInstanceListV2:
      type: array
      items:
        $ref: "#/components/schemas/ReviewInstanceV2"
...
    ErrorReportV2:
      type: object
      required: [message, statusCode, causes]
      properties:
        message:
          type: string
        statusCode:
          type: integer
        causes:
          type: array
          items:
            type: string
...
    BinaryOperatorV2:
      type: string
      enum: ["EQUALS", "NOT_EQUALS", "LESS_THAN", "GREATER_THAN"]
  securitySchemes:
    authorization:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://accounts.google.com/o/oauth2/auth
          scopes:
            openid: open id authorization
            email: email authorization
            profile: profile authorization
    bearerAuth:
      type: http
      scheme: bearer